move masq agent
