+++ pkg/masq/masq.go
@@ -1,6 +1,6 @@
 // SPDX-License-Identifier: APACHE-2.0
 
-package main
+package masq
 
 import (
 	"context"
