+++ pkg/network/ethtool.go
@@ -1,6 +1,6 @@
 // SPDX-License-Identifier: APACHE-2.0
 
-package main
+package network
 
 import (
 	"fmt"
